Fix delete_installed_embedding to reset embedding to initial value
When deleting the last embedding from a library, the embedding field was
being set to an empty list [] instead of the initial "no embedding" value.
This caused subsequent operations like installing new embeddings to fail.

This fix ensures that when all embeddings are deleted, the embedding field
is reset to the initial value:
[{"embedding_status": "no", "embedding_model": "none", "embedding_db": "none",
  "embedded_blocks": 0, "embedding_dims": 0, "time_stamp": "NA"}]

Fixes #1094

Signed-off-by: majiayu000 <1835304752@qq.com>